Fixed roles issue in fileMulter and Cover Image deletion

+++ server/routes/courses.js
@@ -176,12 +176,12 @@ router.post('/info', VerifyToken, async (req, res) => {
 })
 
 router.post('/myCreatedCoursesInfo', VerifyToken, GetRole, async (req, res) => {
+
   try {
     if (req.body.roleID != 1) {
       res.json({ message: "unauthorized" })
       return
     }
-
 
     let courses = []
     const user = await User.findOne({ _id: req.body.userID })
@@ -322,10 +322,14 @@ router.post('/removeFile', VerifyToken, GetRole, async (req, res) => {
       const pathname = course.urlImage.split('/')
       const imageName = pathname[pathname.length - 1]
 
-      const path = 'public/' + req.body.courseID + '/' + imageName
+      // Special case for when first cover image change involves original PEO STRI logo
+      if(pathname[pathname.length - 2] !== 'misc_files')
+      {
+        const path = 'public/' + req.body.courseID + '/' + imageName
 
-      fs.unlinkSync(path)
-      res.json({ 'status': 'module added' });
+        fs.unlinkSync(path)
+      }
+      res.json({ 'status': 'file removed' });
     }
 
   } catch (err) {
